--task=1018843 --user=施立豪 商品发布接口实现

+++ src/main/java/com/example/ehmall/entity/Commodity.java
@@ -69,7 +69,7 @@ public class Commodity implements Serializable {
     /**
      * 原价
      */
-    private int originalprice;
+    private Float originalprice;
     /**
      * 品牌
      */
